refactor xcm_config.rs
extract different types of converters to common_module:
1. CurrencyId <--->MultiLocation
2. AccountId --> MultiLocation
3. AssetId  <---> MultiLocation
